+++ change/@fluentui-react-tag-picker-73122b43-f401-4b38-ba87-5f92358110f5.json
@@ -1,0 +1,7 @@
+{
+  "type": "patch",
+  "comment": "fix: separate headless TagPicker state from styled icon defaults",
+  "packageName": "@fluentui/react-tag-picker",
+  "email": "vgenaev@gmail.com",
+  "dependentChangeType": "patch"
+}
