+++ packages/react-components/react-headless-components-preview/library/bundle-size/TagPicker.fixture.js
@@ -1,0 +1,7 @@
+import * as TagPicker from '@fluentui/react-headless-components-preview/tag-picker';
+
+console.log(TagPicker);
+
+export default {
+  name: '@fluentui/react-headless-components-preview/tag-picker',
+};
