+++ packages/react-components/react-tag-picker/library/src/components/TagPickerControl/TagPickerControl.constants.ts
@@ -1,0 +1,1 @@
+export const tagPickerControlAsideWidthToken = '--fui-TagPickerControl-aside-width' as const;
