+++ packages/react-components/react-tag-picker/library/src/components/TagPickerControl/index.ts
@@ -7,11 +7,12 @@ export type {
   TagPickerControlSlots,
   TagPickerControlState,
 } from './TagPickerControl.types';
+export { tagPickerControlAsideWidthToken } from './TagPickerControl.constants';
 export { renderTagPickerControl_unstable } from './renderTagPickerControl';
-export { useTagPickerControl_unstable, useTagPickerControlBase_unstable } from './useTagPickerControl';
+export { useTagPickerControlBase_unstable } from './useTagPickerControlBase';
+export { useTagPickerControl_unstable } from './useTagPickerControl';
 export {
   iconSizes,
-  tagPickerControlAsideWidthToken,
   tagPickerControlClassNames,
   useTagPickerControlStyles_unstable,
 } from './useTagPickerControlStyles.styles';
